third_party/blink/web_tests/fast/css/invalidation/fieldset-disabled-expected.txt

Use descendant invalidation set for :disabled fieldset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(fieldset, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(fieldset, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
